1.店长系统优化 2.手册系统开发 3.角色系统修改 4.书境系统修改

+++ ZY-APP-SERVER/src/com/zywl/app/server/service/ServerMineService.java
@@ -139,7 +139,7 @@ public class ServerMineService extends BaseService {
                 for (Object o : reward) {
                     JSONObject item = (JSONObject) o;
                     String id = item.getString("id");
-                    int number = item.getIntValue("number");
+                    double number = item.getDoubleValue("number");
                     JSONObject add = new JSONObject();
                     add.put("type", item.getIntValue("type"));
                     add.put("id", id);
